Return shallow copy of array in getCalledActions() (#52)
This fixes a problem where the user might accidentally alter the
calledActions array inside SagaTester. getCalledActions() returned
reference to the array so any function that modifies array in place (like
Array.sort() or Array.reverse()) would modify the internal calledActions
array too. This broke any further calls to getCalledActions() since the
actions in the array were not guaranteed to be in order anymore.